Chapter 9

+++ app/helpers/sessions_helper.rb
@@ -1,11 +1,10 @@
 module SessionsHelper
   def log_in user
-    reset_session
     session[:user_id] = user.id
   end
 
   def current_user
-    @current_user ||= User.find_by(id: session[:user_id])
+    @current_user ||= find_user_from_cookies || find_user_from_session
   end
 
   def logged_in?
@@ -13,7 +12,43 @@ module SessionsHelper
   end
 
   def log_out
-    session.delete(:user_id)
+    forget(current_user) if current_user
+    session.delete(:user_id) if session[:user_id]
+    session.delete(:session_token) if session[:session_token]
     @current_user = nil
+  end
+
+  def forget user
+    user.forget
+    cookies.delete(:user_id)
+    cookies.delete(:remember_token)
+  end
+
+  def remember user
+    user.remember
+    cookies.permanent.signed[:user_id] = user.id
+    cookies.permanent[:remember_token] = user.remember_token
+  end
+
+  private
+
+  def find_user_from_cookies
+    user_id = cookies.signed[:user_id]
+    return unless user_id
+
+    user = User.find_by(id: user_id)
+    return unless user&.authenticated?(cookies[:remember_token])
+
+    log_in(user)
+    user
+  end
+
+  def find_user_from_session
+    user_id = session[:user_id]
+    token = session[:session_token]
+    return unless user_id && token
+
+    user = User.find_by(id: user_id)
+    user if user&.authenticated?(token)
   end
 end
